created better logging method for api calls
