+++ src/applications/check-in/pages/Landing.jsx
@@ -10,7 +10,6 @@ import { receivedAppointmentDetails } from '../actions';
 import { setCurrentToken, clearCurrentSession } from '../utils/session';
 import { createAnalyticsSlug } from '../utils/analytics';
 import { isUUID } from '../utils/token-format-validator';
-import SpeedTracker from '../components/SpeedTracker';
 
 const Landing = props => {
   const { router, setAppointment, location, isLowAuthEnabled } = props;
@@ -33,32 +32,12 @@ const Landing = props => {
       }
 
       if (token) {
-        recordEvent({
-          event: 'api_call',
-          'api-name': 'lorota-token-validation',
-          'api-status': 'started',
-          UUID: token,
-        });
         validateToken(token)
           .then(json => {
             const { data } = json;
             if (data.error || data.errors) {
-              const error = data.error || data.errors;
-              recordEvent({
-                event: 'api_call',
-                'api-name': 'lorota-token-validation',
-                'api-status': 'failed',
-                'error-key': error,
-                UUID: token,
-              });
               goToNextPage(router, URLS.ERROR);
             } else {
-              recordEvent({
-                event: 'api_call',
-                'api-name': 'lorota-token-validation',
-                'api-status': 'success',
-                UUID: token,
-              });
               // dispatch data into redux and local storage
               setAppointment(data, token);
               setCurrentToken(window, token);
@@ -69,15 +48,8 @@ const Landing = props => {
               }
             }
           })
-          .catch(error => {
+          .catch(() => {
             clearCurrentSession(window);
-            recordEvent({
-              event: 'api_call',
-              'api-name': 'lorota-token-validation',
-              'api-status': 'failed',
-              'error-key': error.message || error,
-              UUID: token,
-            });
             goToNextPage(router, URLS.ERROR);
           });
       }
@@ -86,7 +58,6 @@ const Landing = props => {
   );
   return (
     <>
-      <SpeedTracker />
       <LoadingIndicator message="Finding your appointment" />
     </>
   );
